+++ bytecode/src/test/kotlin/io/github/khiaroslav/stringveil/bytecode/SecretLiteralHeuristicsTest.kt
@@ -1,4 +1,4 @@
-package io.github.khiaroslav.stringveil.compiler
+package io.github.khiaroslav.stringveil.bytecode
 
 import kotlin.test.Test
 import kotlin.test.assertNotNull